### 2025-09-26 #9087 @editorialbot
[PRE REVIEW]: MatrixBandwidth.jl: Fast algorithms for matrix bandwidth minimization and recognition
**Submitting author:** @Luis-Varona (<a href="http://orcid.org/0009-0003-7784-5415">Luis M. B. Varona</a>)
**Repository:** https://github.com/Luis-Varona/MatrixBandwidth.jl
**Branch with paper.md** (empty if default branch):
**Version:** v0.2.1
