### 2025-09-26 #9087 @danielskatz
👋 @jbytecode - are you interested in this one?
