### 2025-09-26 c65a537
Replace some URLs with DOIs for JOSS (#170)
This PR replaces some of the URLs in 'joss/paper.bib' with DOIs instead, for the JOSS submission.
