+++ joss/paper.bib
@@ -5,7 +5,7 @@
   eprint = {1906.11786},
   archivePrefix = {arXiv},
   primaryClass = {stat.ML},
-  url = {https://doi.org/10.48550/arXiv.1906.11786}
+  doi = {10.48550/arXiv.1906.11786}
 }
 
 @inproceedings{CM69,
