### 2025-09-27 #9087 @jbytecode
@danielskatz - sure, thank you
