### 2025-09-27 #9087 @jbytecode
@Luis-Varona - Hi, I'm the handling editor for your submission. Thank you for sending your work to the Journal of Open Source Software (JOSS).

We need to assign at least two reviewers for your paper. Do you have any suggestions? You can also look through the list of people at https://reviewers.joss.theoj.org/lookup, which includes individuals who have already volunteered to review for JOSS.

Providing a list of suggestions would be extremely helpful to me. Please do not use the @ character in your suggestions to avoid sending unwanted notifications.

If you don't have any suggestions for suitable reviewers, I will proceed with finding them myself.

Thank you in advance for your help.
