### 2025-09-27 #9087 @Luis-Varona
Hi @jbytecode! Thanks for your help. Here are some suggestions for reviewers:

- ChrisRackauckas (who has talked with me briefly about the package)
- xgdgsc (who has starred the repo)
- matbesancon
- abishektiwari
- fcdimitr
- jhwohlgemuth
- pitmonticone
- sandreza
- dhjx1996
- SunnyXu
- camilochs
- tmigot
- lucaferranti
- ajinkyakadu

Let me know if you need any more.
