### 2025-09-28 #9087 @jbytecode
👋👋👋 Dear @ajinkyakadu  and @lucaferranti 👋👋👋

**[[[PRE REVIEW]: MatrixBandwidth.jl: Fast algorithms for matrix bandwidth minimization and recognition
](https://github.com/openjournals/joss-reviews/issues/9087)**

You can find more information at the top of this Github issue ([LINK](https://github.com/openjournals/joss-reviews/issues/9087)).

Thank you in advance!
