### 2025-09-30 #9087 @lucaferranti
@jbytecode yes I can review this
