### 2025-10-04 Luis-Varona/MatrixBandwidth.jl#171 @tmigot
Make the documentation runable
Connected to https://github.com/openjournals/joss-reviews/issues/9136

The current documentation https://luis-varona.github.io/MatrixBandwidth.jl/dev/ gives a couple of very nice examples. However, I noticed that the output of each line is hardcoded. I think it's nice to have the Julia REPL style, however if something change in one of the algorithms the output might be outdated.

Another solution is to use ```@example``` blocks instead of ```julia-repl```. What do you think?
