### 2025-10-04 Luis-Varona/MatrixBandwidth.jl#172 @tmigot
Community guidelines
Connected to https://github.com/openjournals/joss-reviews/issues/9136

I think this point for the JOSS paper is not currently adressed:

*Are there clear guidelines for third parties wishing to 1. Contribute to the software 2. Report issues or problems with the software 3. Seek support*
