### 2025-10-04 Luis-Varona/MatrixBandwidth.jl#173 @tmigot
Some minor comments on the JOSS paper
Related to https://github.com/openjournals/joss-reviews/issues/9136

This is some small suggestions on the paper that can be further discussed:
- Remove paranthesis: https://github.com/Luis-Varona/MatrixBandwidth.jl/blob/c65a537864d4b1a373ee3c40343d2d8241cb34a2/joss/paper.md?plain=1#L69
- https://github.com/Luis-Varona/MatrixBandwidth.jl/blob/c65a537864d4b1a373ee3c40343d2d8241cb34a2/joss/paper.md?plain=1#L74 Is mentioning the link really necessary given that the JOSS paper is anyway linked to a version of the package?
- Would it be possible to add maybe in one sentence why the package needs an integer programming solver in the deps?
- Add more documentation on the structure returned by the main solvers
- The paper lacks a bit of motivation on why Julia was the language of choice for this project.
- I think it should be also mentioned in the paper that the package BandedMatrices.jl includes symrcm.
