### 2025-10-05 Luis-Varona/MatrixBandwidth.jl#173 @Luis-Varona
Hi, @tmigot! Thanks for the feedback. My responses:

> - [ ] Remove paranthesis: https://github.com/Luis-Varona/MatrixBandwidth.jl/blob/c65a537864d4b1a373ee3c40343d2d8241cb34a2/joss/paper.md?plain=1#L69

Certainly; I'll get right on that.

> - [ ] https://github.com/Luis-Varona/MatrixBandwidth.jl/blob/c65a537864d4b1a373ee3c40343d2d8241cb34a2/joss/paper.md?plain=1#L74 Is mentioning the link really necessary given that the JOSS paper is anyway linked to a version of the package?

Did you, perhaps, mean "mentioning the **version**" rather than "mentioning the **link**"? If so, I was just imitating what [Optim.jl](https://github.com/JuliaNLSolvers/Optim.jl) did in its JOSS paper: "As of version 0.14, the following optimization routines are available." Let me know if you'd like me to remove it regardless.

> - [ ] Would it be possible to add maybe in one sentence why the package needs an integer programming solver in the deps?

Sure&mdash;it's for the Caprara&mdash;Salazar-Gonz&aacute;lez algorithm (in both `MatrixBandwidth.Recognition` and `MatrixBandwidth.Minimization.Exact`). I was a bit hesitant about adding JuMP.jl at first, since it's by far the heaviest dependency, but I figured there was no way around it if I wanted to implement the relevant algorithm's ILP relaxations. In any case, in which section do you think this explanation belongs?

> - [ ] Add more documentation on the structure returned by the main solvers

Again, in which section do you think I should put this? And how exactly do you propose I do so? Enumerate the fields of the structs returned by `minimize_bandwidth` and `has_bandwidth_k_ordering`?

> - [ ] The paper lacks a bit of motivation on why Julia was the language of choice for this project.

Counting the words in the PDF's paper body, I'm already quite close to the 1000-word limit for JOSS submissions (of course, I understand that this is not a _strict_ limit, but still&hellip;). If I explain the need for JuMP.jl and further detail the result structs as discussed above, I'm not sure if I'd have any more space (indeed, the Markdown file already has 1044 words, although the actual rendered PDF of course has less in the body). Further thoughts?

> - [ ] I think it should be also mentioned in the paper that the package BandedMatrices.jl includes symrcm.

I actually didn't clock that&mdash;weird, since I've used BandedMatrices.jl at least once before. Thanks for catching that; I'll certainly add this.
