### 2025-10-05 Luis-Varona/MatrixBandwidth.jl#171 @Luis-Varona
I'll look into this further and get back to you, @tmigot. Thanks again for the feedback. (In the meantime, I've responded to all of your other thoughts/issues 🙂)
