### 2025-10-05 Luis-Varona/MatrixBandwidth.jl#172 @Luis-Varona
On a related note, @tmigot, do you suppose I should remove the section

> To extend the interface with a new matrix bandwidth minimization algorithm, define a new concrete subtype of `AbstractSolver` (or of one of its abstract subtypes like `MetaheuristicSolver`) then implement a corresponding `Minimization._minimize_bandwidth_impl(::AbstractMatrix{Bool}, ::NewSolverType)` method. Similarly, to implement a new bandwidth recognition algorithm, define a new concrete subtype of `AbstractDecider` then implement a corresponding `Recognition._has_bandwidth_k_ordering_impl(::AbstractMatrix{Bool}, ::Integer, ::NewDeciderType)` method. Do *not* attempt to directly implement new `minimize_bandwidth` or `has_bandwidth_k_ordering` methods, as these functions contain common preprocessing logic independent of the specific algorithm used.

from `README.md` (as well as the Documenter-generated website) altogether once `CONTRIBUTING.md` is up? Perhaps just add a very brief `Contributing` section instead, going over very basic rules then referring the reader to `CONTRIBUTING.md`?
