### 2025-10-09 d3c4151
Add JOSS badge to README
This commit adds the 'JOSS: Under Review' badge to the README and Documenter-generated website.
