### 2025-10-09 306cfa5
Add JOSS badge to README (#174)
This commit adds the 'JOSS: Under Review' badge to the README and Documenter-generated website.
