+++ joss/paper.bib
@@ -1,5 +1,6 @@
 @misc{BvMM+19,
-  author = {Balog, Matej and van Merriënboer, Bart and Moitra, Subhodeep and Li, Yujia and Tarlow, Daniel},
+  author = {Balog, Matej and van Merriënboer, Bart and Moitra,
+  Subhodeep and Li, Yujia and Tarlow, Daniel},
   title = {{Fast Training of Sparse Graph Neural Networks on Dense Hardware}},
   year = {2019},
   eprint = {1906.11786},
@@ -48,10 +49,12 @@
 }
 
 @article{GPS76,
-  author = {Gibbs, Norman E. and Poole, Jr., William G. and Stockmeyer, Paul K.},
+  author = {Gibbs, Norman E. and Poole, Jr., William G. and
+  Stockmeyer, Paul K.},
   journal = {SIAM Journal on Numerical Analysis},
   number = {2},
-  title = {{An Algorithm for Reducing the Bandwidth and Profile of a Sparse Matrix}},
+  title = {{An Algorithm for Reducing the Bandwidth and Profile of a
+  Sparse Matrix}},
   volume = {13},
   year = {1976},
   pages = {236--50},
@@ -62,7 +65,8 @@
   author = {Gurari, Eitan M. and Sudborough, Ival Hal},
   journal = {Journal of Algorithms},
   number = {4},
-  title = {{Improved dynamic programming algorithms for bandwidth minimization and the
+  title = {{Improved dynamic programming algorithms for bandwidth
+      minimization and the
   MinCut Linear Arrangement problem}},
   volume = {5},
   year = {1984},
@@ -90,21 +94,45 @@
   doi = {10.1016/j.laa.2024.10.016}
 }
 
+@misc{Jul16,
+  author = {{JuliaLinearAlgebra Developers}},
+  howpublished = {GitHub},
+  title = {{BandedMatrices.jl: A Julia package for representing
+  banded matrices}},
+  year = {2016},
+  url = {https://github.com/JuliaLinearAlgebra/BandedMatrices.jl}
+}
+
 @misc{Krys20,
   author = {Krysl, Petr},
   howpublished = {GitHub},
-  title = {{SymRCM: Reverse Cuthill-McKee node-renumbering algorithm for sparse matrices}},
+  title = {{SymRCM: Reverse Cuthill-McKee node-renumbering algorithm
+  for sparse matrices}},
   year = {2020},
   url = {https://github.com/PetrKryslUCSD/SymRCM.jl}
 }
 
+@article{LDD+23,
+  author = {Lubin, Miles and Dowson, Oscar and Dias Garcia, Joaquim
+  and Huchette, Joey and Legat, Benoît and Vielma, Juan Pablo},
+  journal = {Mathematical Programming Computation},
+  title = {{JuMP 1.0: recent improvements to a modeling language for
+  mathematical optimization}},
+  volume = {15},
+  year = {2023},
+  pages = {581--89},
+  doi = {10.1007/s12532-023-00239-3}
+}
+
 @misc{LLS+01,
-  author = {Lumsdaine, Andrew and Lee, Lie-Quan and Siek, Jeremy G. and Gregor, Doug and
+  author = {Lumsdaine, Andrew and Lee, Lie-Quan and Siek, Jeremy G.
+    and Gregor, Doug and
   McGrath, Kevin D.},
   howpublished = {Boost v1.37.0 documentation},
   title = {{example/cuthill_mckee_ordering.cpp}},
   year = {2001},
-  url = {https://www.boost.org/doc/libs/1_37_0/libs/graph/example/cuthill_mckee_ordering.cpp},
+  url =
+  {https://www.boost.org/doc/libs/1_37_0/libs/graph/example/cuthill_mckee_ordering.cpp},
[…]
