### 2025-10-09 1fbf841
Update JOSS paper (#175)
This commit updates the JOSS paper (currently under review) to address points 1, 3, and 6 of #173 opened by @tmigot. (Points 2, 4, and 5 remain unaddressed and still under discussion in the review process.)
