+++ src/core.jl
@@ -27,42 +27,15 @@ Given an ``n×n`` input matrix, this relatively simple algorithm runs in ``O(n²
 # Examples
 `bandwidth` correctly identifies the bandwidth of a pentadiagonal matrix as ``2`` and does
 not attempt to find a minimizing permutation upon shuffling of its rows and columns:
-```jldoctest
-julia> using Random
-
-julia> Random.seed!(242622);
-
-julia> (n, k) = (8, 2);
-
-julia> perm = randperm(n);
-
-julia> A = (!iszero).(MatrixBandwidth.random_banded_matrix(n, k))
-8×8 BitMatrix:
- 1  0  0  0  0  0  0  0
- 0  1  0  1  0  0  0  0
- 0  0  0  1  1  0  0  0
- 0  1  1  1  0  1  0  0
- 0  0  1  0  0  0  0  0
- 0  0  0  1  0  0  0  0
- 0  0  0  0  0  0  0  0
- 0  0  0  0  0  0  0  0
-
-julia> bandwidth(A)
-2
-
-julia> A_shuffled = A[perm, perm]
-8×8 BitMatrix:
- 0  0  0  0  0  0  1  0
- 0  0  0  0  0  0  0  0
- 0  0  0  1  0  0  0  0
- 0  0  1  0  0  0  1  0
- 0  0  0  0  1  0  0  0
- 0  0  0  0  0  1  1  0
- 1  0  0  1  0  1  1  0
- 0  0  0  0  0  0  0  0
-
-julia> bandwidth(A_shuffled)
-6
+```@repl
+using Random
+Random.seed!(242622);
+(n, k) = (8, 2);
+perm = randperm(n);
+A = (!iszero).(MatrixBandwidth.random_banded_matrix(n, k))
+bandwidth(A)
+A_shuffled = A[perm, perm]
+bandwidth(A_shuffled)
 ```
 
 # Notes
@@ -126,50 +99,23 @@ Given an ``n×n`` input matrix, this relatively simple algorithm runs in ``O(n²
 
 # Examples
 `profile` computes the column profile of a matrix by default:
-```jldoctest
-julia> using Random, SparseArrays
-
-julia> Random.seed!(2287);
-
-julia> (n, p) = (25, 0.05);
-
-julia> A = sprand(n, n, p)
-25×25 SparseMatrixCSC{Float64, Int64} with 29 stored entries:
-⎡⠀⠀⠀⠀⠀⠀⠐⠀⠒⠀⡀⠀⠀⎤
-⎢⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠂⠐⠂⎥
-⎢⠀⠀⠀⢀⠌⠀⠀⠀⢀⠈⠀⠀⠀⎥
-⎢⠠⢄⠁⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⎥
-⎢⠀⠀⠀⠂⡀⠀⠀⠌⠀⠈⠀⠀⠀⎥
-⎢⠀⠀⠀⠀⠀⠀⠢⡀⢄⡈⠀⠀⠀⎥
-⎣⠀⠀⠀⠁⠀⠀⠀⠀⠀⠀⠀⠀⠀⎦
-
-julia> profile(A)
+```@repl
+using Random, SparseArrays
+Random.seed!(2287);
+(n, p) = (25, 0.05);
+A = sprand(n, n, p)
+profile(A)
 211
 ```
 
 The dimension (`:row` or `:col`) can also be explicitly specified:
-```jldoctest
-julia> using Random, SparseArrays
-
-julia> Random.seed!(3647);
-
[…]
