### 2025-10-10 dcbfa6b
Change 'jldoctest' to '@repl' in 'Recognition' submodule
