### 2025-10-11 10b8eea
Change 'jldoctest' to '@repl' in 'Minimization.Exact' submodule
