### 2025-10-11 4875706
Change 'jldoctest' to '@repl' in 'Minimization.Heuristic' submodule
