+++ src/Minimization/Heuristic/solvers/gibbs_poole_stockmeyer.jl
@@ -66,143 +66,44 @@ it hard to verify whether Gibbs–Poole–Stockmeyer finds a truly optimal order
 near-optimal one. Nevertheless, the results are still very good in practice.
 
 Gibbs–Poole–Stockmeyer finds a good ordering for a ``40×40`` matrix:
-```jldoctest
-julia> using Random
-
-julia> Random.seed!(561);
-
-julia> (n, k) = (40, 7);
-
-julia> A = MatrixBandwidth.random_banded_matrix(n, k);
-
-julia> perm = randperm(n);
-
-julia> A_shuffled = A[perm, perm];
-
-julia> bandwidth(A)
-7
-
-julia> bandwidth(A_shuffled)
-37
-
-julia> minimize_bandwidth(A_shuffled, Minimization.GibbsPooleStockmeyer())
-Results of Bandwidth Minimization Algorithm
- * Algorithm: Gibbs–Poole–Stockmeyer
- * Approach: heuristic
- * Minimum Bandwidth: 7
- * Original Bandwidth: 37
- * Matrix Size: 40×40
+```@repl
+using Random
+Random.seed!(561);
+(n, k) = (40, 7);
+A = MatrixBandwidth.random_banded_matrix(n, k);
+perm = randperm(n);
+A_shuffled = A[perm, perm];
+bandwidth(A)
[…]
