### 2025-10-11 d9e4e43
Update changelog to reflect 'jldoctest' -> '@repl' changes
