### 2025-10-11 c06cd0b
Make examples in docstrings runnable (#176)
As suggested by @tmigot in #171, this PR makes the examples in the docstrings runnable by switching from 'jldoctest' to '@repl'. (Previously, the output was hardcoded.) Documentation was built locally and reviewed by myself before the finalization of this PR.

Do note, however, that the examples in 'README.md' (and on the homepage of the Documenter-generated website, with source code at 'docs/srcindex.md') are still hardcoded, since these remain outside of docstrings and thus the '@repl' macro cannot be used.
