+++ src/Minimization/Heuristic/solvers/cuthill_mckee.jl
@@ -52,143 +52,44 @@ it hard to verify whether Cuthill–McKee finds a truly optimal ordering or simp
 near-optimal one. Nevertheless, the results are still very good in practice.
 
 Cuthill–McKee finds a good ordering for a ``30×30`` matrix:
-```jldoctest
-julia> using Random
-
-julia> Random.seed!(13);
-
-julia> (n, k) = (30, 5);
-
-julia> A = MatrixBandwidth.random_banded_matrix(n, k);
-
-julia> perm = randperm(n);
-
-julia> A_shuffled = A[perm, perm];
-
-julia> bandwidth(A)
-5
-
-julia> bandwidth(A_shuffled) # Much larger after shuffling
-25
-
-julia> minimize_bandwidth(A_shuffled, Minimization.CuthillMcKee())
-Results of Bandwidth Minimization Algorithm
- * Algorithm: Cuthill–McKee
- * Approach: heuristic
- * Minimum Bandwidth: 8
- * Original Bandwidth: 25
- * Matrix Size: 30×30
+```@repl
+using Random
+Random.seed!(13);
+(n, k) = (30, 5);
+A = MatrixBandwidth.random_banded_matrix(n, k);
+perm = randperm(n);
+A_shuffled = A[perm, perm];
+bandwidth(A)
[…]
