+++ src/Minimization/core.jl
@@ -40,151 +40,41 @@ than heuristic methods, but even more slowly).
 Certainly, exact solvers will always produce the same optimal bandwidth (but likely
 different orderings):
 
-```jldoctest
-julia> using Random, SparseArrays
-
-julia> Random.seed!(38);
-
-julia> (n, p) = (20, 0.05);
-
-julia> A = sprand(n, n, p);
-
-julia> A = A .+ A' # Ensure structural symmetry
-20×20 SparseMatrixCSC{Float64, Int64} with 36 stored entries:
-⎡⠀⠀⠀⠀⠀⣎⠁⠈⠀⠂⎤
-⎢⠀⠀⠀⠀⡀⠐⠀⠀⠀⠠⎥
-⎢⡠⢤⢀⠈⠠⠂⠀⠀⠔⠀⎥
-⎢⡁⠀⠀⠀⠀⠀⠀⠀⡀⠐⎥
-⎣⠠⠀⠀⡀⠐⠁⢀⠈⢐⠔⎦
-
-julia> res_dcm = minimize_bandwidth(A, Minimization.DelCorsoManzini())
-Results of Bandwidth Minimization Algorithm
- * Algorithm: Del Corso–Manzini
- * Approach: exact
- * Minimum Bandwidth: 3
- * Original Bandwidth: 17
- * Matrix Size: 20×20
-
-julia> A[res_dcm.ordering, res_dcm.ordering]
-20×20 SparseMatrixCSC{Float64, Int64} with 36 stored entries:
-⎡⠠⠂⠤⡀⠀⠀⠀⠀⠀⠀⎤
-⎢⠀⠣⡁⠈⠀⡀⠀⠀⠀⠀⎥
-⎢⠀⠀⠀⠠⡤⠋⠤⡀⠀⠀⎥
-⎢⠀⠀⠀⠀⠀⠣⡀⡨⠢⡀⎥
-⎣⠀⠀⠀⠀⠀⠀⠈⠢⠁⠀⎦
-
-julia> res_sgs = minimize_bandwidth(A, Minimization.SaxeGurariSudborough())
[…]
