### 2025-10-11 421579c
Update JOSS paper
this commit updates the JOSS paper (currently under review) to address point 2 of issue #173 by removing the mention of the package's version. (Points 4 and 5 remain unaddressed and will be tackled next.)
