### 2025-10-14 Luis-Varona/MatrixBandwidth.jl#173 @tmigot
Good idea
