### 2025-10-14 Luis-Varona/MatrixBandwidth.jl#171 @tmigot
Ahaha Don't forget to enjoy the poutine as well 😀
