### 2025-10-20 3b0b461
Add unit test for 'README' example (#178)
This commit adds a unit test to ensure that the example code blocks in 'README.md' and 'docs/src/index.md' align with the actual output of the package on 64-bit architectures. (The exact orderings generated on 32-bit architectures may differ, although as verified by many other unit tests, they still produce valid results.)
