+++ test/readme_example.jl
@@ -1,0 +1,82 @@
+# Copyright 2025 Luis M. B. Varona
+#
+# Licensed under the MIT license <LICENSE or
+# http://opensource.org/licenses/MIT>. This file may not be copied, modified, or
+# distributed except according to those terms.
+
+"""
+    ReadmeExample
+
+Unit test to ensure that the example code blocks in `README.md` and `docs/src/index.md`
+(which corresponds to the homepage of the Documenter.jl-generated documentation) align with
+the actual output of the package on 64-bit architectures. (The exact orderings generated on
+32-bit architectures may differ, although as verified by many other unit tests, they still
+produce valid results.)
+
+Instead of testing that the REPL display output is the same, we instead test that all the
+`AbstractResult` fields (including the orderings that produced the sparse matrices shown in
+the example blocks when the code was first run) match.
+"""
+module ReadmeExample
+
+using MatrixBandwidth
+using Random
+using SparseArrays
+using Test
+
+# We avoid wrapping these long lines by turning off formatting
+#! format: off
+const EXAMPLE_MATRIX_A = sparse([52, 35, 36, 10, 26, 52, 30, 49, 9, 22, 49, 9, 26, 52, 27, 30, 5, 7, 3, 12, 15, 10, 36, 29, 30, 31, 50, 51, 10, 35, 42, 18, 36, 49, 6, 27, 53, 55, 37, 3, 7, 46, 8, 23, 34, 13, 55, 4, 8, 13, 54, 13, 28, 49, 2, 16, 2, 12, 18, 25, 54, 42, 17, 41, 48, 60, 48, 49, 26, 57, 43, 44, 4, 6, 21, 34, 45, 13, 13, 1, 3, 7, 23, 30, 38, 24, 29, 59, 47, 59, 56, 57, 43], [1, 2, 2, 3, 3, 3, 4, 4, 5, 6, 6, 7, 7, 7, 8, 8, 9, 9, 10, 10, 10, 12, 12, 13, 13, 13, 13, 13, 15, 16, 17, 18, 18, 21, 22, 23, 23, 24, 25, 26, 26, 26, 27, 27, 28, 29, 29, 30, 30, 30, 30, 31, 34, 34, 35, 35, 36, 36, 36, 37, 38, 41, 42, 42, 43, 43, 44, 45, 46, 47, 48, 48, 49, 49, 49, 49, 49, 50, 51, 52, 52, 52, 53, 54, 54, 55, 55, 56, 57, 57, 59, 59, 60], [0.24576738978057033, 0.9815652517640157, 0.037958122629155766, 0.676385808781116, 0.22478555599426664, 0.7573202446419934, 0.23688023076089482, 0.783637414278683, 0.8919509435876607, 0.3639787152009696, 0.11048593562719855, 0.9718860273776642, 0.1517761364629282, 0.4789284822576675, 0.27648464173344767, 0.8913857951102127, 0.8919509435876607, 0.9718860273776642, 0.676385808781116, 0.659803854423641, 0.8659788467336631, 0.659803854423641, 0.9331834434144175, 0.6686777203645599, 0.11165753599399597, 0.07395734815915611, 0.17159724883213667, 0.0633887493611911, 0.8659788467336631, 0.9936028846837545, 0.7294297925493887, 0.9173898474530946, 0.766493941664756, 0.6903237814160842, 0.3639787152009696, 0.43860615108882994, 0.7382944526610568, 0.7593
[…]
