+++ CHANGELOG.md
@@ -12,6 +12,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 
 ### Changed
 
+- Added a unit test in `test/readme_example.jl` to ensure that the example code blocks in `README.md` and `docs/src/index.md` (which corresponds to the homepage of the Documenter.jl-generated documentation) align with the actual output of the package on 64-bit architectures (#178).
 - Made examples in the docstrings runnable (thus avoiding hardcoding output) by switching from `jldoctest` to `@repl` (#176).
 
 ## [0.2.1] - 2025-09-24
