### 2025-10-20 bd3cdd9
Add Julia 1.12 to CI
This commit updates the GitHub Actions CI workflow to include Julia 1.12 in the testing matrix.

We also move the 'CHANGELOG.md' entry for PR #178 from the 'Changed' section to the 'Added' section - its initial placement in the 'Changed' section was just a careless oversight.
