### 2025-10-20 #9136 @jbytecode
@Luis-Varona - Thank you for clarifying this. I hope JOSS review finishes before the acceptance and latest proof-readings of your other paper and you can properly cite your published JOSS submission in your work. I think there is nothing to do right now.
