### 2025-10-22 54316be
Link to docs on main names in 'README.md' (#180)
This commit adds clear links to the documentation on 'minimize_bandwidth', 'has_bandwidth_k_ordering', and their respective output structs in 'README.md' and 'docs/src/index.md'.

(This should address point 4 raised by @tmigot in issue #173.)
