### 2025-10-24 Luis-Varona/MatrixBandwidth.jl#173 @Luis-Varona
@tmigot, the last point (about "why Julia?") should be addressed tomorrow, hopefully. 🙂 After that, on to the `CONTRIBUTING.md`! Do you have any other general things you need me to fix in the repository?
