+++ .github/workflows/draft-pdf.yml
@@ -21,7 +21,7 @@ jobs:
           journal: joss
           paper-path: joss/paper.md
       - name: Upload
-        uses: actions/upload-artifact@v4
+        uses: actions/upload-artifact@v5
         with:
           name: paper
           path: joss/paper.pdf
