### 2025-10-27 2ddfa79
Add justification for Julia to JOSS paper
This commit adds a brief sentence to the Summary section of the JOSS paper explaining why Julia was chosen as the implementation language for this project. Accordingly, this should close issue #173, pending approval from @tmigot.
