### 2025-10-27 Luis-Varona/MatrixBandwidth.jl#173 @Luis-Varona
Hi @tmigot, I rewrote the **Summary** section of the JOSS paper:

> &hellip; offers fast algorithms for matrix bandwidth minimization and recognition. Julia's combination of easy syntax and high performance, along with its rapidly growing ecosystem for scientific computing, made it the ideal language of choice for this project.

Do you think this is sufficient to close this issue? (If so, I'll merge #182 and close this.)
