### 2025-10-31 Luis-Varona/MatrixBandwidth.jl#173 @Luis-Varona
Hi @tmigot, just following up! 🙂 Anything else besides the `CONTRIBUTING.md` you'd like me to fix after merging #182 (assuming it's satisfactory to you)?
