### 2025-10-31 Luis-Varona/MatrixBandwidth.jl#173 @tmigot
Yes, it sounds good to me, thanks
