### 2025-10-31 b410f7d
Add justification for Julia to JOSS paper (#182)
This commit adds a brief sentence to the Summary section of the JOSS paper explaining why Julia was chosen as the implementation language for this project. (Resolves #173.)
