### 2025-11-11 Luis-Varona/MatrixBandwidth.jl#172 @Luis-Varona
Hi @tmigot, sorry for the late follow-up. I'll try to get the `CONTRIBUTING.md` done tomorrow, and hopefully it'll meet your expectations. Besides this issue, is there anything else you'd like me to change for your JOSS review?
