### 2025-11-13 #9136 @jbytecode
@danielskatz - Hi Daniel, thank you for pinging. It looks like some issues are still open in the target repository.

@tmigot - The author states that they made changes as you requested. Could you please check them and update your status?

@lucaferranti - May I kindly ask for a status update?

Thank you in advance.
