### 2025-11-16 c19c940
Add a 'CONTRIBUTING.md'
We add a 'CONTRIBUTING.md' with contributing guidelines. (Resolves #139 and #172.)
