+++ CONTRIBUTING.md
@@ -1,0 +1,95 @@
+# Contributing
+
+Welcome to MatrixBandwidth.jl! Thank you for your interest in contributing to this package. Kindly peruse these guidelines before opening an issue or submitting a pull request (PR).
+
+## Issues
+
+Most issues are either bug reports or feature requests; we describe the expectations for each type below. If your issue does not fit into either category, feel free to open it anyway&mdash;just be clear, concise, and of course respectful!
+
+### Bug reports
+
+If you encounter a bug, please include a [minimal reproducible example](https://en.wikipedia.org/wiki/Minimal_reproducible_example) containing the code necessary to reproduce the bug, along with the expected and actual behavior. Additionally, please provide information about your operating system, Julia version, and MatrixBandwidth.jl version. (If you are unsure about how to find this information, feel free to reach out anyway, and we can guide you through the process.)
+
+### Feature requests
+
+When suggesting a new feature, please provide a clear and concise description of your idea, along with concrete motivation for its inclusion in the package. If applicable, include references to relevant literature or existing implementations in other software. And of course, if you wish to implement the feature yourself, please indicate so in your issue, and we can get you underway to submitting a PR!
+
+## Pull requests
+
+Before writing a PR, we would prefer if you first reached out via an issue to check whether your contribution would fit well with the repository. Once in the process of actually writing a PR, please adhere to the following guidelines:
+
+### Correctness
+
+When introducing new features to the package, unit tests are of paramount importance to ensure that your code behaves as expected. Please add them to the [`test/`](test/) directory, following the existing structure and conventions: a unit test meant for a method in `src/<path>/<file>.jl` should be placed in `test/<path>/<file>.jl`.
+
+### Documentation
+
+Equally crucial is proper documentation. Judicious use of inline comments in complex sections is encouraged to enhance code readability, regardless of whether the method/struct is public or private. For public methods/structs, please also include docstrings matching the style of existing documentation. At bare minimum, methods should have a one-liner summary of their purpose along with descriptions of all arguments and return values, and structs should have a one-liner summary of their purpose along with descriptions of all fields. Less trivial methods/structs should have more extensive documentation&mdash;examples in particular are highly recommended.
+
+### Style
+
+Format your code with [JuliaFormatter](https://github.com/domluna/JuliaFormatter.jl) in the package's environment before pushing; our [`.JuliaFormatter.toml`](.JuliaFormatter.toml) file will automatically enforce Invenia's [BlueStyle](https://github.com/invenia/BlueStyle), with a few modifications.
+
+### Efficiency
+
+While not always necessary, strive to write efficient code, especially for performance-critical sections. Moreover, try to avoid adding unncessary dependencies to the package; if you absolutely *must*, please justify its inclusion in your PR description.
+
+### New algorithms
+
[…]
