### 2025-11-16 #9136 @tmigot
Hi @jbytecode @Luis-Varona
I have completed my review. The only remaining open point concerns the community guidelines, but I am confident that @Luis-Varona will provide a satisfactory fix. Aside from that, I am happy to recommend the paper for acceptance.
