### 2025-11-16 f50f4bb
Add a 'CONTRIBUTING.md' (#183)
We add a 'CONTRIBUTING.md' with contributing guidelines.
