### 2025-11-16 1e07a4c
Fix indexing conflict in Documenter website
The bump from Documenter v1.15.0 to v1.16.0 resulted in errors building the documentation due to conflicting headers between the private and public API sections. This fixes that bug.
