### 2025-11-16 de9f48e
Fix indexing conflict in Documenter website (#184)
The bump from Documenter v1.15.0 to v1.16.0 resulted in errors building the documentation due to conflicting headers between the private and public API sections. This fixes that bug.
