+++ docs/src/public_api.md
@@ -9,42 +9,42 @@ Documentation for [MatrixBandwidth](https://github.com/Luis-Varona/MatrixBandwid
 !!! note
     The following documentation covers only the public API of the package. For internal details, see the [private API documentation](private_api.md).
 
-## `MatrixBandwidth`
+## `MatrixBandwidth` (Public)
 
 ```@autodocs
 Modules = [MatrixBandwidth]
 Private = false
 ```
 
-## `MatrixBandwidth.Minimization`
+## `MatrixBandwidth.Minimization` (Public)
 
 ```@autodocs
 Modules = [MatrixBandwidth.Minimization]
 Private = false
 ```
 
-## `MatrixBandwidth.Minimization.Exact`
+## `MatrixBandwidth.Minimization.Exact` (Public)
 
 ```@autodocs
 Modules = [MatrixBandwidth.Minimization.Exact]
 Private = false
 ```
 
-## `MatrixBandwidth.Minimization.Heuristic`
+## `MatrixBandwidth.Minimization.Heuristic` (Public)
 
 ```@autodocs
 Modules = [MatrixBandwidth.Minimization.Heuristic]
 Private = false
 ```
 
-## `MatrixBandwidth.Minimization.Metaheuristic`
+## `MatrixBandwidth.Minimization.Metaheuristic` (Public)
 
 ```@autodocs
 Modules = [MatrixBandwidth.Minimization.Metaheuristic]
 Private = false
 ```
 
-## `MatrixBandwidth.Recognition`
+## `MatrixBandwidth.Recognition` (Public)
 
 ```@autodocs
 Modules = [MatrixBandwidth.Recognition]
