### 2025-11-16 4e72ef4
Prepare metadata for v0.2.2
