+++ README.md
@@ -7,7 +7,7 @@
   <tr>
     <td>Metadata</td>
     <td>
-      <img src="https://img.shields.io/badge/version-v0.2.1-pink.svg" alt="Version">
+      <img src="https://img.shields.io/badge/version-v0.2.2-pink.svg" alt="Version">
       <a href="https://opensource.org/licenses/MIT"><img src="https://img.shields.io/badge/License-MIT-A31F34.svg" alt="License: MIT"></a>
       <a href="https://github.com/JuliaDiff/BlueStyle"><img src="https://img.shields.io/badge/code%20style-blue-4495d1.svg" alt="Code Style: Blue"></a>
       <a href="https://joss.theoj.org/papers/9677739548a4f05ab398b427df1c78b9"><img src="https://joss.theoj.org/papers/9677739548a4f05ab398b427df1c78b9/status.svg" alt="JOSS: Under Review"></a>
@@ -319,6 +319,6 @@ We welcome all bug reports, feature requests, and contributions! Please refer to
 
 ## Project status
 
-The latest stable release of MatrixBandwidth.jl is v0.2.1. Although several metaheuristic algorithms are still under development, the rest of the package is fully functional and covered by unit tests.
+The latest stable release of MatrixBandwidth.jl is v0.2.2. Although several metaheuristic algorithms are still under development, the rest of the package is fully functional and covered by unit tests.
 
 Currently, MatrixBandwidth.jl's core functions generically accept any input of the type `AbstractMatrix{<:Number}`, not behaving any differently when given sparsely stored matrices (e.g., from the [SparseArrays.jl](https://github.com/JuliaSparse/SparseArrays.jl) standard library package). Capabilities for directly handling graph inputs (aiming to reduce the matrix bandwidth of a graph's adjacency) are also not available. Given that bandwidth reduction is often applied to sparse matrices and graphs, these limitations will be addressed in a future release of the package.
