### 2025-11-16 #9136 @Luis-Varona
Hi @jbytecode - @tmigot has finished reviewing the paper, I believe. Just waiting on approval from @lucaferranti, then I hope to bump the package to v0.2.2 before generating the final JOSS paper (which I believe is linked to a package version, right?). 🙂
