### 2025-11-17 5a4f9ff
Update README.md and index.md (#185)
We move the 'Algorithms' section in the documentation further down and explicitly import with 'using MatrixBandwidth' in the 'Basic use' examples, pursuant to @lucaferranti's suggestions.
