### 2025-11-30 #9136 @Luis-Varona
Hi @lucaferranti, thanks again for the feedback! Some clarifications I have:

> yes, the readme is good, my recommendation would be to factor out the getting started section in its own "tutorial" section for better navigation. This also leads to

May I ask what you would have me keep in the "Basic use" section, and where I'd put the rest of it? You advise removing it entirely from the README, if I understand correctly; so it'd just show up somewhere on the Documenter-generated website, not the GitHub repo directly?

> Yes, I think it would be cool to have an "Applications" section in the documentation showcasing in practice the usage of your library. If for example you have a research paper where you use this library, the companion code of that would be the great. This would refer to the part in parenthesis of the review bullet point

I don't have an entire research paper yet where I've used it, although I'm in the process of drafting one (but I haven't coded stuff up just yet). If ever, would this stuff go directly in the JOSS paper, or again just somewhere in the documentation?
