### 2025-11-30 #9136 @lucaferranti
> I don't have an entire research paper yet where I've used it, although I'm in the process of drafting one (but I haven't coded stuff up just yet). If ever, would this stuff go directly in the JOSS paper, or again just somewhere in the documentation?

I would say in the documentation, my reasoning is that JOSS paper is the snapshot but the software keep evolving and improving. If (when :) ) you have an application paper I think the code will make a nice addition to the documentation, if there's none yet, no point in forcing it, research takes time and the documentation is very good already :)
