### 2025-11-30 #9136 @lucaferranti
(sorry for the two separate messages, accidentally pressed enter too early 😅 )

>  if I understand correctly; so it'd just show up somewhere on the Documenter-generated website, not the GitHub repo directly?

I think the readme doesn't need changes, it's very good already.

I was referring that this page: https://github.com/Luis-Varona/MatrixBandwidth.jl/blob/main/docs/src/index.md could be factored into two parts, keeping the home-page minimal and moving the already existing basic examples into its own tutorial page. But again, this is not a required change, if you think the current structure is better, that is ok of course
