### 2025-11-30 #9136 @Luis-Varona
Hi @lucaferranti, thank you! So just to confirm:

1. Since there's no actual research paper done yet (although I'm working on a draft, so I'll update with the suggested changes in the future), no need to add any concrete examples yet?
2. Leave the README alone and split up the Documenter-generated homepage to have a separate **Tutorial** section with more detail than the current **Basic use**? Or am I misunderstanding?

Thanks again 🙂
