### 2025-12-08 ef9a6ba
Update docs for JOSS
In the Documenter-generated website, we factor out the Basic use examples into a separate Tutorial page, as per @lucaferranti's request.
