### 2025-12-08 b89ff7b
Update docs for JOSS (#187)
In the Documenter-generated website, we factor out the Basic use examples into a separate Tutorial page, as per @lucaferranti's request.
