+++ docs/src/tutorial.md
@@ -1,0 +1,201 @@
+```@meta
+CurrentModule = MatrixBandwidth
+```
+
+# Tutorial
+
+MatrixBandwidth.jl offers unified interfaces for both bandwidth minimization and bandwidth recognition via the `minimize_bandwidth` and `has_bandwidth_k_ordering` functions, respectively—the algorithm itself is specified as an argument. Comprehensive documentation for these two methods, as well as their output structs, is available in the [public API documentation](public_api.md). We go over some basic examples below.
+
+## Getting started
+
+Throughout these examples, we shall work with a random sparse, structurally symmetric matrix of size $60 \times 60$ with initial bandwidth $51$:
+
+```julia-repl
+julia> using Random, SparseArrays
+
+julia> Random.seed!(8675309);
+
+julia> A = sprand(60, 60, 0.01); A = A + A' # Ensure structural symmetry
+60×60 SparseMatrixCSC{Float64, Int64} with 93 stored entries:
+⎡⠀⠀⠀⠀⠠⠀⠀⠀⠀⠀⠀⠀⠠⠀⢀⠀⠀⠒⠀⠀⠀⠀⠀⠀⡀⠨⠀⠀⠀⠀⎤
+⎢⠀⠀⠀⠀⠅⠀⠀⠀⠀⠀⠐⠀⠠⡀⢀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠂⠠⠀⠀⠀⠀⎥
+⎢⠀⠂⠁⠁⢀⠐⠀⠂⠀⠀⠀⠀⠀⠀⠀⠀⠀⢀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⎥
+⎢⠀⠀⠀⠀⠠⠀⠀⠀⠀⠀⠀⠀⠀⠀⠉⠁⠀⡀⠀⠀⠀⠀⠀⠀⠈⠁⠀⠀⠀⠀⎥
+⎢⠀⠀⠀⠀⠀⠀⠀⠀⠐⠀⠀⠀⠀⠀⠀⠀⠀⠐⠀⠀⠈⠀⠀⠀⠀⠀⠀⠀⠀⠀⎥
[…]
