+++ test/Project.toml
@@ -7,6 +7,9 @@ Random = "9a3f8284-a2c9-5f02-9a11-845980a1fd5c"
 SparseArrays = "2f01184e-e22b-5df5-ae63-d93ebab69eaf"
 Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
 
+[sources]
+MatrixBandwidth = {path = ".."}
+
 [compat]
 Aqua = "0.8.5"
 Graphs = "1.10"
