+++ docs/Project.toml
@@ -7,6 +7,9 @@ MatrixBandwidth = "075aa41b-24fd-4573-b25f-92cae432c0f8"
 Random = "9a3f8284-a2c9-5f02-9a11-845980a1fd5c"
 SparseArrays = "2f01184e-e22b-5df5-ae63-d93ebab69eaf"
 
+[sources]
+MatrixBandwidth = {path = ".."}
+
 [compat]
 Combinatorics = "1.0"
 Documenter = "1.3"
