### 2025-12-08 cdc5aaf
Fix docs build update from #187 (#189)
