### 2025-12-08 4b0fced
Update date of JOSS paper (#190)
We update the date of the JOSS paper to 7 December 2025, awaiting approval and publication.

---------

Co-authored-by: Luis-Varona <Luis-Varona@users.noreply.github.com>
