### 2025-12-08 95dfa8f
Prepare metadata for v0.2.2
