+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 
 ## [Unreleased]
 
+## [0.2.2] - 2025-12-08
+
 ### Added
 
 - Added a `CONTRIBUTING.md` file with contributing guidelines (#183).
@@ -126,7 +128,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 
 - Released the initial stable version of the package.
 
-[unreleased]: https://github.com/Luis-Varona/MatrixBandwidth.jl/compare/v0.2.1...HEAD
+[unreleased]: https://github.com/Luis-Varona/MatrixBandwidth.jl/compare/v0.2.2...HEAD
+[0.2.2]: https://github.com/Luis-Varona/MatrixBandwidth.jl/releases/tag/v0.2.2
 [0.2.1]: https://github.com/Luis-Varona/MatrixBandwidth.jl/releases/tag/v0.2.1
 [0.2.0]: https://github.com/Luis-Varona/MatrixBandwidth.jl/releases/tag/v0.2.0
 [0.1.4]: https://github.com/Luis-Varona/MatrixBandwidth.jl/releases/tag/v0.1.4
