### 2025-12-08 ad5ba65
Prepare metadata for v0.2.2 (#191)
