+++ Project.toml
@@ -8,7 +8,7 @@ homepage = "https://luis-varona.github.io/MatrixBandwidth.jl/"
 maintainers = ["Luis M. B. Varona <lm.varona@outlook.com>"]
 readme = "README.md"
 repository = "https://github.com/Luis-Varona/MatrixBandwidth.jl"
-version = "0.2.1"
+version = "0.2.2"
 
 [deps]
 Combinatorics = "861a8166-3701-5b0c-9a16-15d98fcdc6aa"
