+++ v0.2.2/search_index.js
@@ -1,0 +1,3 @@
+var documenterSearchIndex = {"docs":
+[{"location":"public_api/#Public-API-Documentation","page":"Public API","title":"Public API Documentation","text":"Documentation for MatrixBandwidth's public API.\n\nnote: Note\nThe following documentation covers only the public API of the package. For internal details, see the private API documentation.","category":"section"},{"location":"public_api/#MatrixBandwidth-(Public)","page":"Public API","title":"MatrixBandwidth (Public)","text":"","category":"section"},{"location":"public_api/#MatrixBandwidth.Minimization-(Public)","page":"Public API","title":"MatrixBandwidth.Minimization (Public)","text":"","category":"section"},{"location":"public_api/#MatrixBandwidth.Minimization.Exact-(Public)","page":"Public API","title":"MatrixBandwidth.Minimization.Exact (Public)","text":"","category":"section"},{"location":"public_api/#MatrixBandwidth.Minimization.Heuristic-(Public)","page":"Public API","title":"MatrixBandwidth.Minimization.Heuristic (Public)","text":"","category":"section"},{"location":"public_api/#MatrixBandwidth.Minimization.Metaheuristic-(Public)","page":"Public API","title":"MatrixBandwidth.Minimization.Metaheuristic (Public)","text":"","category":"section"},{"location":"public_api/#MatrixBandwidth.Recognition-(Public)","page":"Public API","title":"MatrixBandwidth.Recognition (Public)","text":"","category":"section"},{"location":"public_api/#References","page":"Public API","title":"References","text":"A. Caprara and J.-J. Salazar-González. Laying Out Sparse Graphs with Provably Minimum Bandwidth. INFORMS Journal on Computing 17, 356–73 (2005).\n\n\n\nW. M. Chan and A. George. A linear time implementation of the reverse Cuthill–McKee algorithm. BIT Numerical Mathematics 20, 8–14 (1980).\n\n\n\nE. Cuthill and J. McKee. Reducing the bandwidth of sparse symmetric matrices. In: Proceedings of the 24th National Conference of the ACM (Brandon Systems Press, 1969); pp. 157–72.\n\n\n\nG. M. Del Corso and G. Manzini. Finding Exact Solutions to the Bandwidth Minimization Problem. Computing 62, 189–203 (1999).\n\n\n\nM. R. Garey, R. L. Graham, D. S. Johnson and D. E. Knuth. Complexity Results for Bandwidth Minimization. SIAM Journal on Applied Mathematics 34, 477–95 (1978).\n\n\n\nJ. A. George. Computer Implementation of the Finite Element Method. Ph.D. Thesis, Department of Computer Science, Stanford University (1971).\n\n\n\nN. E. Gibbs, W. G. Poole Jr. and P. K. Stockmeyer. An Algorithm for Reducing the Bandwidth and Profile of a Sparse Matrix. SIAM Journal on Numerical Analysis 13, 236–50 (1976).\n\n\n\nE. M. Gurari and I. H. Sudborough. Improved dynamic programming algorithms for bandwidth minimization and the   MinCut Linear Arrangement problem. Journal of Algorithms 5, 531–46 (1984).\n\n\n\nJ. G. Lewis. Implementation of the Gibbs–Poole–Stockmeyer and Gibbs–King Algorithms. ACM Transactions on Mathematical Software 8, 180–89 (1982).\n\n\n\nL. O. Mafteiu-Scai. The Bandwidths of a Matrix. A Survey of Algorithms. Annals of West University of Timisoara - Mathematics and Computer Science 52, 183–223 (2014).\n\n\n\nJ. B. Saxe. Dynamic-Programming Algorithms for Recognizin
[…]
