### 2025-12-08 264a19b
minor fixes in paper and bibtex
