+++ joss/paper.bib
@@ -116,7 +116,7 @@
   author = {Lubin, Miles and Dowson, Oscar and Dias Garcia, Joaquim
   and Huchette, Joey and Legat, Benoît and Vielma, Juan Pablo},
   journal = {Mathematical Programming Computation},
-  title = {{JuMP 1.0: recent improvements to a modeling language for
+  title = {{JuMP 1.0: {R}ecent improvements to a modeling language for
   mathematical optimization}},
   volume = {15},
   year = {2023},
@@ -187,4 +187,19 @@
   \emph{S}-bandwidth of an undirected graph}},
   year = {2025},
   url = {https://github.com/GraphQuantum/SDiagonalizability.jl}
+}
+
+
+@article{julia,
+  doi       = {10.1137/141000671},
+  url       = {https://doi.org/10.1137%2F141000671},
+  year      = 2017,
+  month     = {jan},
+  publisher = {Society for Industrial {\&} Applied Mathematics ({SIAM})},
+  volume    = {59},
+  number    = {1},
+  pages     = {65--98},
+  author    = {Jeff Bezanson and Alan Edelman and Stefan Karpinski and Viral B. Shah},
+  title     = {{J}ulia: A Fresh Approach to Numerical Computing},
+  journal   = {{SIAM} Review}
 }
