### 2025-12-08 f570a65
Add Julia citation to JOSS paper; update date (#194)
Add Julia citation to JOSS paper; update date

---------

Co-authored-by: Luis-Varona <Luis-Varona@users.noreply.github.com>
