+++ joss/paper.bib
@@ -1,3 +1,15 @@
+@article{BEKS17,
+  author = {Bezanson, Jeff and Edelman, Alan and Karpinski, Stefan
+  and Shah, Viral B.},
+  journal = {SIAM Review},
+  number = {1},
+  title = {{Julia: A Fresh Approach to Numerical Computing}},
+  volume = {59},
+  year = {2017},
+  pages = {65--98},
+  doi = {10.1137/141000671}
+}
+
 @misc{BvMM+19,
   author = {Balog, Matej and van Merriënboer, Bart and Moitra,
   Subhodeep and Li, Yujia and Tarlow, Daniel},
