### 2025-12-09 #9136 @Luis-Varona
Hi @jbytecode, I just released v0.2.2 yesterday, and the only thing changed since then was adding the Julia citation to the JOSS paper. Would you still like me to release v0.2.3 (which would involve no changes to the code or to the documentation)?
