### 2025-12-09 #9136 @Luis-Varona
1. I have released v0.2.2 (both on GitHub and to the Julia general registry).
2. The DOI is [10.5281/zenodo.17872329](https://doi.org/10.5281/zenodo.17872329), and as mentioned, the tagged release number is v0.2.2.
3. The title, author (myself), and ORCID (0009-0003-7784-5415) match. (I will note, however, that the affiliations don't&mdash;the JOSS paper lists my departments at my university, and the Zenodo archive just lists the university itself.)
5. The Zenodo archive indeed lists the MIT License.
All done, @jbytecode!
