### 2025-12-10 #9136 @jbytecode
@Luis-Varona - There are some errors in the final compilation stage and it's possibly about naming/referencing figures. Could you please take a look at the paper and check if you properly named/cited figures and images within the manuscript? Thank you in advance.
