### 2025-12-11 #9136 @jbytecode
@Luis-Varona - Due to the bug found in pandoc, we need to refactor the paper's source code as stated above.
