### 2025-12-11 #9136 @tarleb
FYI: The pandoc bug is tracked here: https://github.com/jgm/pandoc/issues/11342
