### 2025-12-11 501d049
Fix figure rendering due to pandoc bug (#195)
Due to a bug in pandoc (tracked by jgm/pandoc#11342), neither minipage nor subfigure renders properly. This attempts to replicate the old behavior using a tabular environment instead.

---------

Co-authored-by: Luis-Varona <Luis-Varona@users.noreply.github.com>
