### 2025-12-11 e0dda17
Update paper.bib
