+++ joss/paper.bib
@@ -3,7 +3,7 @@
   and Shah, Viral B.},
   journal = {SIAM Review},
   number = {1},
-  title = {{Julia: A Fresh Approach to Numerical Computing}},
+  title = {{J}ulia: A Fresh Approach to Numerical Computing},
   volume = {59},
   year = {2017},
   pages = {65--98},
@@ -13,7 +13,7 @@
 @misc{BvMM+19,
   author = {Balog, Matej and van Merriënboer, Bart and Moitra,
   Subhodeep and Li, Yujia and Tarlow, Daniel},
-  title = {{Fast Training of Sparse Graph Neural Networks on Dense Hardware}},
+  title = {Fast Training of Sparse Graph Neural Networks on Dense Hardware},
   year = {2019},
   eprint = {1906.11786},
   archivePrefix = {arXiv},
@@ -24,7 +24,7 @@
 @inproceedings{CM69,
   author = {Cuthill, E. and McKee, J.},
   booktitle = {Proceedings of the 24th National Conference of the ACM},
-  title = {{Reducing the bandwidth of sparse symmetric matrices}},
+  title = {Reducing the bandwidth of sparse symmetric matrices},
   year = {1969},
   pages = {157--72},
   publisher = {Brandon Systems Press},
@@ -35,7 +35,7 @@
   author = {Caprara, Alberto and Salazar-González, Juan-José},
   journal = {INFORMS Journal on Computing},
   number = {3},
-  title = {{Laying Out Sparse Graphs with Provably Minimum Bandwidth}},
+  title = {Laying Out Sparse Graphs with Provably Minimum Bandwidth},
   volume = {17},
   year = {2005},
   pages = {356-73},
@@ -45,7 +45,7 @@
 @article{DM99,
   author = {Del Corso, G. M. and Manzini, G.},
   journal = {Computing},
-  title = {{Finding Exact Solutions to the Bandwidth Minimization Problem}},
+  title = {Finding Exact Solutions to the Bandwidth Minimization Problem},
   volume = {62},
   year = {1999},
   pages = {189--203},
@@ -55,7 +55,7 @@
 @phdthesis{Geo71,
   author = {George, J. Alan},
   school = {Department of Computer Science, Stanford University},
-  title = {{Computer Implementation of the Finite Element Method}},
+  title = {Computer Implementation of the Finite Element Method},
   year = {1971},
   url = {https://apps.dtic.mil/sti/tr/pdf/AD0726171.pdf}
 }
@@ -65,8 +65,8 @@
   Stockmeyer, Paul K.},
   journal = {SIAM Journal on Numerical Analysis},
   number = {2},
-  title = {{An Algorithm for Reducing the Bandwidth and Profile of a
-  Sparse Matrix}},
+  title = {An Algorithm for Reducing the Bandwidth and Profile of a
+  Sparse Matrix},
   volume = {13},
   year = {1976},
   pages = {236--50},
@@ -77,9 +77,9 @@
   author = {Gurari, Eitan M. and Sudborough, Ival Hal},
   journal = {Journal of Algorithms},
   number = {4},
-  title = {{Improved dynamic programming algorithms for bandwidth
+  title = {Improved dynamic programming algorithms for bandwidth
       minimization and the
-  MinCut Linear Arrangement problem}},
+  {MinCut Linear Arrangement} problem},
   volume = {5},
   year = {1984},
   pages = {531--46},
@@ -89,7 +89,7 @@
 @article{JMP25,
   author = {Johnston, Nathaniel and Moein, Shirin and Plosker, Sarah},
   journal = {Linear Algebra and its Applications},
[…]
