### 2025-12-11 580a59d
updates for JOSS publication (#196)
* Update paper.md

* Update paper.bib
