### 2025-12-11 bc5a1c0
Reformat joss/paper.bib with tex-fmt
