+++ joss/paper.bib
@@ -78,8 +78,7 @@
   journal = {Journal of Algorithms},
   number = {4},
   title = {Improved dynamic programming algorithms for bandwidth
-      minimization and the
-  {MinCut Linear Arrangement} problem},
+  minimization and the {MinCut Linear Arrangement} problem},
   volume = {5},
   year = {1984},
   pages = {531--46},
@@ -138,8 +137,7 @@
 
 @misc{LLS+01,
   author = {Lumsdaine, Andrew and Lee, Lie-Quan and Siek, Jeremy G.
-    and Gregor, Doug and
-  McGrath, Kevin D.},
+  and Gregor, Doug and McGrath, Kevin D.},
   howpublished = {Boost v1.37.0 documentation},
   title = {example/cuthill_mckee_ordering.cpp},
   year = {2001},
@@ -184,8 +182,7 @@
   journal = {SIAM Journal on Algebraic and Discrete Methods},
   number = {4},
   title = {Dynamic-Programming Algorithms for Recognizing
-      Small-Bandwidth Graphs in
-  Polynomial Time},
+  Small-Bandwidth Graphs in Polynomial Time},
   volume = {1},
   pages = {363--69},
   year = {1980},
@@ -195,8 +192,8 @@
 @misc{VJP25,
   author = {Varona, Luis M. B. and Johnston, Nathaniel and Plosker, Sarah},
   howpublished = {GitHub},
-  title = {SDiagonalizability: A dynamic algorithm to minimize or recognize the
-  \emph{{S}}-bandwidth of an undirected graph},
+  title = {SDiagonalizability: A dynamic algorithm to minimize or
+  recognize the \emph{{S}}-bandwidth of an undirected graph},
   year = {2025},
   url = {https://github.com/GraphQuantum/SDiagonalizability.jl}
 }
