### 2025-12-11 b160f58
Update JOSS paper/biblio for publication (#197)
We add the country to the author affiliations in the JOSS paper, and modify slightly the casing in the bibliography.

---------

Co-authored-by: Daniel S. Katz <d.katz@ieee.org>
Co-authored-by: Luis-Varona <Luis-Varona@users.noreply.github.com>
