### 2025-12-11 #9136 @danielskatz
> [@danielskatz](https://github.com/danielskatz) - The official paper page says "404 File not found"; it appears to be a problem with only mine, though.
>
> https://joss.theoj.org/papers/10.21105/joss.09136

This is typically a network caching issue, where you tried to see the file before it was available, and then that 404 gets cached. It should go away in a few hours, and the file will work for you.  It does work for me currently.
