### 2025-12-11 #9136 @Luis-Varona
> Congratulations to [@Luis-Varona](https://github.com/Luis-Varona) (Luis M. B. Varona) on your JOSS publication!!
>
> And thanks to [@lucaferranti](https://github.com/lucaferranti) & [@tmigot](https://github.com/tmigot) for reviewing and to [@jbytecode](https://github.com/jbytecode) for editing! JOSS completely depends on volunteers, and we couldn't do this without you

Thanks again, all!
